requirements.txt
toml
paho-mqtt
pexpect
requests
json-rpc
python-mpd2
